removes okgroup
