+++ cmd/apichecker/apicheck.go
@@ -33,7 +33,7 @@ const (
 	github               = "GitHub Sha Check"
 	htmlScrape           = "HTML String Check"
 	pathBinance          = "https://binance-docs.github.io/apidocs/spot/en/#change-log"
-	pathOkCoin           = "https://www.okcoin.com/docs/en/#change-change"
+	pathOkCoin           = "https://www.o.com/docs/en/#change-change"
 	pathFTX              = "https://github.com/ftexchange/ftx"
 	pathBTSE             = "https://www.btse.com/apiexplorer/spot/#btse-spot-api"
 	pathBitfinex         = "https://docs.bitfinex.com/docs/changelog"
