+++ exchanges/okcoin/okcoin_types.go
@@ -1,19 +1,14 @@
-package okgroup
+package okcoin
 
 import (
-	"time"
-
+	"errors"
 	"github.com/thrasher-corp/gocryptotrader/currency"
 	"github.com/thrasher-corp/gocryptotrader/exchanges/asset"
+	"sync"
+	"time"
 )
 
-// Order types
-const (
-	NormalOrder = iota
-	PostOnlyOrder
-	FillOrKillOrder
-	ImmediateOrCancelOrder
-)
+var errNoAccountDepositAddress = errors.New("no account deposit address")
 
 // PerpSwapInstrumentData stores instrument data for perpetual swap contracts
 type PerpSwapInstrumentData struct {
@@ -106,9 +101,11 @@ type PerpSwapFundingRates struct {
 type GetAccountCurrenciesResponse struct {
 	Name          string  `json:"name"`
 	Currency      string  `json:"currency"`
-	CanDeposit    int     `json:"can_deposit,string"`
-	CanWithdraw   int     `json:"can_withdraw,string"`
-	MinWithdrawal float64 `json:"min_withdrawal,string"`
+	Chain         string  `json:"chain"`
+	CanInternal   bool    `json:"can_internal,string"`
+	CanWithdraw   bool    `json:"can_withdraw,string"`
+	CanDeposit    bool    `json:"can_deposit,string"`
+	MinWithdrawal float64 `json:"min_withdrawal"`
 }
 
 // WalletInformationResponse response data for WalletInformation
@@ -1354,7 +1351,7 @@ type GetETTSettlementPriceHistoryResponse struct {
 	Price float64 `json:"price"`
 }
 
-// OrderStatus Holds OKGroup order status values
+// OrderStatus Holds Okcoin order status values
 var OrderStatus = map[int64]string{
 	-3: "pending cancel",
 	-2: "cancelled",
@@ -1516,4 +1513,295 @@ type WebsocketErrorResponse struct {
 	Event     string `json:"event"`
 	Message   string `json:"message"`
 	ErrorCode int64  `json:"errorCode"`
+}
+
+// List of all websocket channels to subscribe to
+const (
+	// Orderbook events
+	okcoinWsOrderbookUpdate  = "update"
+	okcoinWsOrderbookPartial = "partial"
+	// API subsections
+	okcoinWsSwapSubsection    = "swap/"
+	okcoinWsIndexSubsection   = "index/"
+	okcoinWsFuturesSubsection = "futures/"
+	okcoinWsSpotSubsection    = "spot/"
+	// Shared API endpoints
+	okcoinWsCandle         = "candle"
+	okcoinWsCandle60s      = okcoinWsCandle + "60s"
+	okcoinWsCandle180s     = okcoinWsCandle + "180s"
+	okcoinWsCandle300s     = okcoinWsCandle + "300s"
+	okcoinWsCandle900s     = okcoinWsCandle + "900s"
+	okcoinWsCandle1800s    = okcoinWsCandle + "1800s"
+	okcoinWsCandle3600s    = okcoinWsCandle + "3600s"
+	okcoinWsCandle7200s    = okcoinWsCandle + "7200s"
+	okcoinWsCandle14400s   = okcoinWsCandle + "14400s"
+	okcoinWsCandle21600s   = okcoinWsCandle + "21600"
+	okcoinWsCandle43200s   = okcoinWsCandle + "43200s"
+	okcoinWsCandle86400s   = okcoinWsCandle + "86400s"
+	okcoinWsCandle604900s  = okcoinWsCandle + "604800s"
+	okcoinWsTicker         = "ticker"
+	okcoinWsTrade          = "trade"
+	okcoinWsDepth          = "depth"
+	okcoinWsDepth5         = "depth5"
+	okcoinWsAccount        = "account"
+	okcoinWsMarginAccount  = "margin_account"
+	okcoinWsOrder          = "order"
+	okcoinWsFundingRate    = "funding_rate"
+	okcoinWsPriceRange     = "price_range"
+	okcoinWsMarkPrice      = "mark_price"
+	okcoinWsPosition       = "position"
+	okcoinWsEstimatedPrice = "estimated_price"
+	// Spot endpoints
+	okcoinWsSpotTicker        = okcoinWsSpotSubsection + okcoinWsTicker
+	okcoinWsSpotCandle60s     = okcoinWsSpotSubsection + okcoinWsCandle60s
+	okcoinWsSpotCandle180s    = okcoinWsSpotSubsection + okcoinWsCandle180s
+	okcoinWsSpotCandle300s    = okcoinWsSpotSubsection + okcoinWsCandle300s
+	okcoinWsSpotCandle900s    = okcoinWsSpotSubsection + okcoinWsCandle900s
+	okcoinWsSpotCandle1800s   = okcoinWsSpotSubsection + okcoinWsCandle1800s
+	okcoinWsSpotCandle3600s   = okcoinWsSpotSubsection + okcoinWsCandle3600s
+	okcoinWsSpotCandle7200s   = okcoinWsSpotSubsection + okcoinWsCandle7200s
+	okcoinWsSpotCandle14400s  = okcoinWsSpotSubsection + okcoinWsCandle14400s
+	okcoinWsSpotCandle21600s  = okcoinWsSpotSubsection + okcoinWsCandle21600s
+	okcoinWsSpotCandle43200s  = okcoinWsSpotSubsection + okcoinWsCandle43200s
+	okcoinWsSpotCandle86400s  = okcoinWsSpotSubsection + okcoinWsCandle86400s
+	okcoinWsSpotCandle604900s = okcoinWsSpotSubsection + okcoinWsCandle604900s
+	okcoinWsSpotTrade         = okcoinWsSpotSubsection + okcoinWsTrade
+	okcoinWsSpotDepth         = okcoinWsSpotSubsection + okcoinWsDepth
+	okcoinWsSpotDepth5        = okcoinWsSpotSubsection + okcoinWsDepth5
+	okcoinWsSpotAccount       = okcoinWsSpotSubsection + okcoinWsAccount
+	okcoinWsSpotMarginAccount = okcoinWsSpotSubsection + okcoinWsMarginAccount
+	okcoinWsSpotOrder         = okcoinWsSpotSubsection + okcoinWsOrder
+	// Swap endpoints
+	okcoinWsSwapTicker        = okcoinWsSwapSubsection + okcoinWsTicker
+	okcoinWsSwapCandle60s     = okcoinWsSwapSubsection + okcoinWsCandle60s
+	okcoinWsSwapCandle180s    = okcoinWsSwapSubsection + okcoinWsCandle180s
+	okcoinWsSwapCandle300s    = okcoinWsSwapSubsection + okcoinWsCandle300s
+	okcoinWsSwapCandle900s    = okcoinWsSwapSubsection + okcoinWsCandle900s
+	okcoinWsSwapCandle1800s   = okcoinWsSwapSubsection + okcoinWsCandle1800s
+	okcoinWsSwapCandle3600s   = okcoinWsSwapSubsection + okcoinWsCandle3600s
+	okcoinWsSwapCandle7200s   = okcoinWsSwapSubsection + okcoinWsCandle7200s
+	okcoinWsSwapCandle14400s  = okcoinWsSwapSubsection + okcoinWsCandle14400s
+	okcoinWsSwapCandle21600s  = okcoinWsSwapSubsection + okcoinWsCandle21600s
+	okcoinWsSwapCandle43200s  = okcoinWsSwapSubsection + okcoinWsCandle43200s
+	okcoinWsSwapCandle86400s  = okcoinWsSwapSubsection + okcoinWsCandle86400s
+	okcoinWsSwapCandle604900s = okcoinWsSwapSubsection + okcoinWsCandle604900s
+	okcoinWsSwapTrade         = okcoinWsSwapSubsection + okcoinWsTrade
+	okcoinWsSwapDepth         = okcoinWsSwapSubsection + okcoinWsDepth
+	okcoinWsSwapDepth5        = okcoinWsSwapSubsection + okcoinWsDepth5
+	okcoinWsSwapFundingRate   = okcoinWsSwapSubsection + okcoinWsFundingRate
+	okcoinWsSwapPriceRange    = okcoinWsSwapSubsection + okcoinWsPriceRange
+	okcoinWsSwapMarkPrice     = okcoinWsSwapSubsection + okcoinWsMarkPrice
+	okcoinWsSwapPosition      = okcoinWsSwapSubsection + okcoinWsPosition
+	okcoinWsSwapAccount       = okcoinWsSwapSubsection + okcoinWsAccount
+	okcoinWsSwapOrder         = okcoinWsSwapSubsection + okcoinWsOrder
+	// Index endpoints
+	okcoinWsIndexTicker        = okcoinWsIndexSubsection + okcoinWsTicker
+	okcoinWsIndexCandle60s     = okcoinWsIndexSubsection + okcoinWsCandle60s
+	okcoinWsIndexCandle180s    = okcoinWsIndexSubsection + okcoinWsCandle180s
+	okcoinWsIndexCandle300s    = okcoinWsIndexSubsection + okcoinWsCandle300s
+	okcoinWsIndexCandle900s    = okcoinWsIndexSubsection + okcoinWsCandle900s
+	okcoinWsIndexCandle1800s   = okcoinWsIndexSubsection + okcoinWsCandle1800s
+	okcoinWsIndexCandle3600s   = okcoinWsIndexSubsection + okcoinWsCandle3600s
+	okcoinWsIndexCandle7200s   = okcoinWsIndexSubsection + okcoinWsCandle7200s
+	okcoinWsIndexCandle14400s  = okcoinWsIndexSubsection + okcoinWsCandle14400s
+	okcoinWsIndexCandle21600s  = okcoinWsIndexSubsection + okcoinWsCandle21600s
+	okcoinWsIndexCandle43200s  = okcoinWsIndexSubsection + okcoinWsCandle43200s
+	okcoinWsIndexCandle86400s  = okcoinWsIndexSubsection + okcoinWsCandle86400s
+	okcoinWsIndexCandle604900s = okcoinWsIndexSubsection + okcoinWsCandle604900s
+	// Futures endpoints
+	okcoinWsFuturesTicker         = okcoinWsFuturesSubsection + okcoinWsTicker
+	okcoinWsFuturesCandle60s      = okcoinWsFuturesSubsection + okcoinWsCandle60s
+	okcoinWsFuturesCandle180s     = okcoinWsFuturesSubsection + okcoinWsCandle180s
+	okcoinWsFuturesCandle300s     = okcoinWsFuturesSubsection + okcoinWsCandle300s
+	okcoinWsFuturesCandle900s     = okcoinWsFuturesSubsection + okcoinWsCandle900s
+	okcoinWsFuturesCandle1800s    = okcoinWsFuturesSubsection + okcoinWsCandle1800s
+	okcoinWsFuturesCandle3600s    = okcoinWsFuturesSubsection + okcoinWsCandle3600s
+	okcoinWsFuturesCandle7200s    = okcoinWsFuturesSubsection + okcoinWsCandle7200s
+	okcoinWsFuturesCandle14400s   = okcoinWsFuturesSubsection + okcoinWsCandle14400s
+	okcoinWsFuturesCandle21600s   = okcoinWsFuturesSubsection + okcoinWsCandle21600s
+	okcoinWsFuturesCandle43200s   = okcoinWsFuturesSubsection + okcoinWsCandle43200s
+	okcoinWsFuturesCandle86400s   = okcoinWsFuturesSubsection + okcoinWsCandle86400s
+	okcoinWsFuturesCandle604900s  = okcoinWsFuturesSubsection + okcoinWsCandle604900s
+	okcoinWsFuturesTrade          = okcoinWsFuturesSubsection + okcoinWsTrade
+	okcoinWsFuturesEstimatedPrice = okcoinWsFuturesSubsection + okcoinWsTrade
+	okcoinWsFuturesPriceRange     = okcoinWsFuturesSubsection + okcoinWsPriceRange
+	okcoinWsFuturesDepth          = okcoinWsFuturesSubsection + okcoinWsDepth
+	okcoinWsFuturesDepth5         = okcoinWsFuturesSubsection + okcoinWsDepth5
+	okcoinWsFuturesMarkPrice      = okcoinWsFuturesSubsection + okcoinWsMarkPrice
+	okcoinWsFuturesAccount        = okcoinWsFuturesSubsection + okcoinWsAccount
+	okcoinWsFuturesPosition       = okcoinWsFuturesSubsection + okcoinWsPosition
+	okcoinWsFuturesOrder          = okcoinWsFuturesSubsection + okcoinWsOrder
+
+	okcoinWsRateLimit = 30
+
+	allowableIterations = 25
+	delimiterColon      = ":"
+	delimiterDash       = "-"
+
+	maxConnByteLen = 4096
+)
+
+// orderbookMutex Ensures if two entries arrive at once, only one can be
+// processed at a time
+var orderbookMutex sync.Mutex
+
+var defaultSpotSubscribedChannels = []string{okcoinWsSpotDepth,
+	okcoinWsSpotCandle300s,
+	okcoinWsSpotTicker,
+	okcoinWsSpotTrade}
+
+var defaultFuturesSubscribedChannels = []string{okcoinWsFuturesDepth,
+	okcoinWsFuturesCandle300s,
+	okcoinWsFuturesTicker,
+	okcoinWsFuturesTrade}
+
+var defaultIndexSubscribedChannels = []string{okcoinWsIndexCandle300s,
+	okcoinWsIndexTicker}
+
+var defaultSwapSubscribedChannels = []string{okcoinWsSwapDepth,
+	okcoinWsSwapCandle300s,
+	okcoinWsSwapTicker,
+	okcoinWsSwapTrade,
+	okcoinWsSwapFundingRate,
+	okcoinWsSwapMarkPrice}
+
+// SetErrorDefaults sets the full error default list
+func (o *OKCoin) SetErrorDefaults() {
+	o.ErrorCodes = map[string]error{
+		"0":     errors.New("successful"),
+		"1":     errors.New("invalid parameter in url normally"),
+		"30001": errors.New("request header \"OK_ACCESS_KEY\" cannot be blank"),
+		"30002": errors.New("request header \"OK_ACCESS_SIGN\" cannot be blank"),
+		"30003": errors.New("request header \"OK_ACCESS_TIMESTAMP\" cannot be blank"),
+		"30004": errors.New("request header \"OK_ACCESS_PASSPHRASE\" cannot be blank"),
+		"30005": errors.New("invalid OK_ACCESS_TIMESTAMP"),
+		"30006": errors.New("invalid OK_ACCESS_KEY"),
+		"30007": errors.New("invalid Content_Type, please use \"application/json\" format"),
+		"30008": errors.New("timestamp request expired"),
+		"30009": errors.New("system error"),
+		"30010": errors.New("api validation failed"),
+		"30011": errors.New("invalid IP"),
+		"30012": errors.New("invalid authorization"),
+		"30013": errors.New("invalid sign"),
+		"30014": errors.New("request too frequent"),
+		"30015": errors.New("request header \"OK_ACCESS_PASSPHRASE\" incorrect"),
+		"30016": errors.New("you are using v1 apiKey, please use v1 endpoint. If you would like to use v3 endpoint, please subscribe to v3 apiKey"),
+		"30017": errors.New("apikey's broker id does not match"),
+		"30018": errors.New("apikey's domain does not match"),
+		"30020": errors.New("body cannot be blank"),
+		"30021": errors.New("json data format error"),
+		"30023": errors.New("required parameter cannot be blank"),
+		"30024": errors.New("parameter value error"),
+		"30025": errors.New("parameter category error"),
+		"30026": errors.New("requested too frequent; endpoint limit exceeded"),
+		"30027": errors.New("login failure"),
+		"30028": errors.New("unauthorized execution"),
+		"30029": errors.New("account suspended"),
+		"30030": errors.New("endpoint request failed. Please try again"),
+		"30031": errors.New("token does not exist"),
+		"30032": errors.New("pair does not exist"),
+		"30033": errors.New("exchange domain does not exist"),
+		"30034": errors.New("exchange ID does not exist"),
+		"30035": errors.New("trading is not supported in this website"),
+		"30036": errors.New("no relevant data"),
+		"30037": errors.New("endpoint is offline or unavailable"),
+		"30038": errors.New("user does not exist"),
+		"32001": errors.New("futures account suspended"),
+		"32002": errors.New("futures account does not exist"),
+		"32003": errors.New("canceling, please wait"),
+		"32004": errors.New("you have no unfilled orders"),
+		"32005": errors.New("max order quantity"),
+		"32006": errors.New("the order price or trigger price exceeds USD 1 million"),
+		"32007": errors.New("leverage level must be the same for orders on the same side of the contract"),
+		"32008": errors.New("max. positions to open (cross margin)"),
+		"32009": errors.New("max. positions to open (fixed margin)"),
+		"32010": errors.New("leverage cannot be changed with open positions"),
+		"32011": errors.New("futures status error"),
+		"32012": errors.New("futures order update error"),
+		"32013": errors.New("token type is blank"),
+		"32014": errors.New("your number of contracts closing is larger than the number of contracts available"),
+		"32015": errors.New("margin ratio is lower than 100% before opening positions"),
+		"32016": errors.New("margin ratio is lower than 100% after opening position"),
+		"32017": errors.New("no BBO"),
+		"32018": errors.New("the order quantity is less than 1, please try again"),
+		"32019": errors.New("the order price deviates from the price of the previous minute by more than 3%"),
+		"32020": errors.New("the price is not in the range of the price limit"),
+		"32021": errors.New("leverage error"),
+		"32022": errors.New("this function is not supported in your country or region according to the regulations"),
+		"32023": errors.New("this account has outstanding loan"),
+		"32024": errors.New("order cannot be placed during delivery"),
+		"32025": errors.New("order cannot be placed during settlement"),
+		"32026": errors.New("your account is restricted from opening positions"),
+		"32027": errors.New("cancelled over 20 orders"),
+		"32028": errors.New("account is suspended and liquidated"),
+		"32029": errors.New("order info does not exist"),
+		"33001": errors.New("margin account for this pair is not enabled yet"),
+		"33002": errors.New("margin account for this pair is suspended"),
+		"33003": errors.New("no loan balance"),
+		"33004": errors.New("loan amount cannot be smaller than the minimum limit"),
+		"33005": errors.New("repayment amount must exceed 0"),
+		"33006": errors.New("loan order not found"),
+		"33007": errors.New("status not found"),
+		"33008": errors.New("loan amount cannot exceed the maximum limit"),
+		"33009": errors.New("user ID is blank"),
+		"33010": errors.New("you cannot cancel an order during session 2 of call auction"),
+		"33011": errors.New("no new market data"),
+		"33012": errors.New("order cancellation failed"),
+		"33013": errors.New("order placement failed"),
+		"33014": errors.New("order does not exist"),
+		"33015": errors.New("exceeded maximum limit"),
+		"33016": errors.New("margin trading is not open for this token"),
+		"33017": errors.New("insufficient balance"),
+		"33018": errors.New("this parameter must be smaller than 1"),
+		"33020": errors.New("request not supported"),
+		"33021": errors.New("token and the pair do not match"),
+		"33022": errors.New("pair and the order do not match"),
+		"33023": errors.New("you can only place market orders during call auction"),
+		"33024": errors.New("trading amount too small"),
+		"33025": errors.New("base token amount is blank"),
+		"33026": errors.New("transaction completed"),
+		"33027": errors.New("cancelled order or order cancelling"),
+		"33028": errors.New("the decimal places of the trading price exceeded the limit"),
+		"33029": errors.New("the decimal places of the trading size exceeded the limit"),
+		"34001": errors.New("withdrawal suspended"),
+		"34002": errors.New("please add a withdrawal address"),
+		"34003": errors.New("sorry, this token cannot be withdrawn to xx at the moment"),
+		"34004": errors.New("withdrawal fee is smaller than minimum limit"),
+		"34005": errors.New("withdrawal fee exceeds the maximum limit"),
+		"34006": errors.New("withdrawal amount is lower than the minimum limit"),
+		"34007": errors.New("withdrawal amount exceeds the maximum limit"),
+		"34008": errors.New("insufficient balance"),
+		"34009": errors.New("your withdrawal amount exceeds the daily limit"),
+		"34010": errors.New("transfer amount must be larger than 0"),
+		"34011": errors.New("conditions not met"),
+		"34012": errors.New("the minimum withdrawal amount for NEO is 1, and the amount must be an integer"),
+		"34013": errors.New("please transfer"),
+		"34014": errors.New("transfer limited"),
+		"34015": errors.New("subaccount does not exist"),
+		"34016": errors.New("transfer suspended"),
+		"34017": errors.New("account suspended"),
+		"34018": errors.New("incorrect trades password"),
+		"34019": errors.New("please bind your email before withdrawal"),
+		"34020": errors.New("please bind your funds password before withdrawal"),
+		"34021": errors.New("not verified address"),
+		"34022": errors.New("withdrawals are not available for sub accounts"),
+		"35001": errors.New("contract subscribing does not exist"),
+		"35002": errors.New("contract is being settled"),
+		"35003": errors.New("contract is being paused"),
+		"35004": errors.New("pending contract settlement"),
+		"35005": errors.New("perpetual swap trading is not enabled"),
+		"35008": errors.New("margin ratio too low when placing order"),
+		"35010": errors.New("closing position size larger than available size"),
+		"35012": errors.New("placing an order with less than 1 contract"),
+		"35014": errors.New("order size is not in acceptable range"),
+		"35015": errors.New("leverage level unavailable"),
+		"35017": errors.New("changing leverage level"),
+		"35019": errors.New("order size exceeds limit"),
+		"35020": errors.New("order price exceeds limit"),
+		"35021": errors.New("order size exceeds limit of the current tier"),
+		"35022": errors.New("contract is paused or closed"),
+		"35030": errors.New("place multiple orders"),
+		"35031": errors.New("cancel multiple orders"),
+		"35061": errors.New("invalid instrument_id"),
+	}
 }
